Moved ResourceDefinitionExample from GettingStarted to JsonApiDotNetCoreExample
