+++ src/Examples/JsonApiDotNetCoreExample/Controllers/ModelsController.cs
@@ -1,9 +1,10 @@
 using JsonApiDotNetCore.Configuration;
 using JsonApiDotNetCore.Controllers;
 using JsonApiDotNetCore.Services;
+using JsonApiDotNetCoreExample.Models;
 using Microsoft.Extensions.Logging;
 
-namespace GettingStarted.ResourceDefinitionExample
+namespace JsonApiDotNetCoreExample.Controllers
 {
     public sealed class ModelsController : JsonApiController<Model>
     {
